refactoring form
