+++ src/components/NFTMinter/index.js
@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import UploadAndDisplayImage from 'components/ImageUplaod/imageUpload';
+import Form from '../Form/form';
 import 'components/NFTMinter/styles.css';
 
 const NFTMinter = ({props}) => {
@@ -7,17 +7,7 @@ const NFTMinter = ({props}) => {
   return (
       <div id="minter">
           NFT Minter
-            <form>
-            Properties of your NFT
-            <br></br>
-            <label htmlFor='nftname'> Name of the NFT</label>
-            <input type="text" name="nftname" id="nftname"></input>
-            <br></br>
-            <label htmlFor='nftdescription'> Description</label>
-            <input type="text" name="nftdescription" id="nftdescription"></input>
-            <br></br>
-            <UploadAndDisplayImage/>
-            </form>
+            <Form/>
       </div>
   );
 }
